look for either of 10.4u or 10.5 SDK, just like portconfigure
git-svn-id: https://svn.macports.org/repository/macports/trunk/base@33053 d073be05-634f-4543-b044-5fe20cf6d1d6

+++ src/port1.0/portmain.tcl
@@ -106,7 +106,8 @@ if {[info exists variations(macosx)] && $variations(macosx) == "+"} {
             return -code error "Default universal variant only works with ports based on configure"
         }
         eval configure.args-append ${configure.universal_args}
-        if {![file exists /Developer/SDKs/MacOSX10.4u.sdk/]} {
+        if {![file exists /Developer/SDKs/MacOSX10.4u.sdk] &&
+            ![file exists /Developer/SDKs/MacOSX10.5.sdk]} {
             return -code error "MacOS X 10.4 universal SDK is not installed (are we running on 10.3? did you forget to install it?) and building with +universal will very likely fail"
         }
         eval configure.cflags-append ${configure.universal_cflags}
